+++ src/Evaluation/B_Spline/test_2d_2.py
@@ -47,7 +47,7 @@ def main():
                   [3.00, -2.50], 
                   [3.75, -1.25], 
                   [4.00,  0.75], 
-                  [5.00,  1.00]], dtype=np.float32)
+                  [5.00,  1.00]], dtype=np.float64)
 
     # Initialization of a specific class to work with B-Spline curves from input 
     # control points.
