+++ src/Evaluation/Bezier/Bernstein_Polynomials.py
@@ -40,7 +40,7 @@ def main():
                   [3.00, -2.50], 
                   [3.75, -1.25], 
                   [4.00,  0.75], 
-                  [5.00,  1.00]], dtype=np.float32)
+                  [5.00,  1.00]], dtype=np.float64)
     
     # The value of the time must be within the interval: 
     #   0.0 <= x <= 1.0
