+++ src/Evaluation/Bezier/test_3d_1.py
@@ -49,7 +49,7 @@ def main():
                   [3.00, -2.50,  1.00], 
                   [3.75, -1.25, -0.50], 
                   [4.00,  0.75,  1.50], 
-                  [5.00,  1.00, -1.50]], dtype=np.float32)
+                  [5.00,  1.00, -1.50]], dtype=np.float64)
 
     # Initialization of a specific class to work with Bézier curves.
     B_Cls = Bezier.Bezier_Cls(CONST_BEZIER_CURVE['method'], P,
